requirements.txt
# Job scheduling library (for timed tasks, cron-style jobs)
APScheduler

# Core Discord API wrapper
hikari

# Command framework built on Hikari (slash & prefix commands)
hikari-lightbulb

# Load environment variables from a .env file into os.environ
python-dotenv

# MongoDB client (sync)
pymongo

# Friendly datetime library used by APScheduler
pendulum

# Clash of Clans API wrapper
coc.py

Pillow